refac: file structure
