+++ src/components/pages/MainPage/components/AnimatedBar/index.tsx
