+++ src/components/pages/MainPage/components/AnimatedBar/styled.ts
