+++ src/components/pages/MainPage/components/Works/index.tsx
@@ -1,8 +1,7 @@
-import Image from 'next/future/image';
 import { RefObject, useEffect, useRef, useState } from 'react';
-import styled from 'styled-components';
 import { nextImageLoader } from 'utils/imageLoader';
-import { ARCHIVE_DATA } from './pages/ArchivePage/constants';
+import { ARCHIVE_DATA } from '../../../ArchivePage/constants';
+import { Work, WorksDiv } from './styled';
 
 interface Props {
   unit: RefObject<HTMLDivElement>;
@@ -66,48 +65,5 @@ const Works: React.FC<Props> = (props) => {
     </WorksDiv>
   );
 };
-
-const WorksDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  overflow-x: hidden;
-  overflow-y: scroll;
-  & > div {
-    width: 100%;
-    height: 100%;
-    & img {
-      opacity: 0.5;
-    }
-  }
-  @media (min-width: 1024px) {
-    gap: 100px;
-    min-width: 1950px;
-  }
-  @media (max-width: 1023px) {
-    gap: 400px;
-    & > div {
-      margin-top: -50px;
-      display: flex;
-      min-height: 100vh;
-      justify-content: center;
-      transform: rotate(-15deg);
-    }
-    width: 100%;
-    height: 100vh;
-    background: #000;
-    z-index: 0;
-  }
-`;
-
-const Work = styled(Image)`
-  @media (min-width: 1024px) {
-    width: 1800px;
-  }
-  @media (max-width: 1023px) {
-    height: 115%;
-  }
-`;
 
 export default Works;
