+++ src/components/templates/PortfolioLayout/components/Content/styled.ts
